Makes general component BuildList, creates userSets fragment
