+++ client/graphql/fragments/userSets.graphql
@@ -1,0 +1,40 @@
+fragment userSets on User {
+customSets(first: $first, after: $after, filters: $filters) {
+      edges {
+        node {
+          id
+          defaultClass {
+            id
+            name
+            faceImageUrl
+          }
+          name
+          level
+          equippedItems {
+            id
+            slot {
+              id
+              order
+            }
+            item {
+              id
+              imageUrl
+            }
+          }
+          tagAssociations {
+            id
+            associationDate
+            customSetTag {
+              id
+              name
+              imageUrl
+            }
+          }
+        }
+      }
+      pageInfo {
+        hasNextPage
+        endCursor
+      }
+    }
+    }
