+++ client/graphql/queries/userProfileSets.graphql
@@ -1,3 +1,5 @@
+#import "graphql/fragments/userSets.graphql"
+
 query userProfileSets(
   $username: String!
   $first: Int!
@@ -7,43 +9,6 @@ query userProfileSets(
   userByName(username: $username) {
     username
     id
-    customSets(first: $first, after: $after, filters: $filters) {
-      edges {
-        node {
-          id
-          name
-          level
-          defaultClass {
-            id
-            name
-            faceImageUrl
-          }
-          equippedItems {
-            id
-            slot {
-              id
-              order
-            }
-            item {
-              id
-              imageUrl
-            }
-          }
-          tagAssociations {
-            id
-            associationDate
-            customSetTag {
-              id
-              name
-              imageUrl
-            }
-          }
-        }
-      }
-      pageInfo {
-        hasNextPage
-        endCursor
-      }
-    }
+    ...userSets
   }
 }
